requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.1
pillow==11.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
tk==0.1.0
